earthcandy logic wip

+++ FlagQuizSwiftUI/FlagQuizSwiftUI/View/MyPage/MyPageViewModel.swift
@@ -36,12 +36,13 @@ final class MyPageViewModel: ObservableObject {
     }
     
     func deleteAccount() async {
+        guard let userId = user?.id else { return }
         do {
             try container.services.authService.signOut()
-            let userId = try await container.services.authService.deleteAccount()
             try await container.services.userService.deleteUser(of: userId)
             try await container.services.earthCandyService.deleteEarthCandy(ofUser: userId)
             try await container.services.frogService.deleteFrog(ofUser: userId)
+            _ = try await container.services.authService.deleteAccount()
         } catch {
             debugPrint(error.localizedDescription)
         }
